requirements.txt
attrs==23.1.0
certifi==2023.7.22
exceptiongroup==1.1.3
greenlet==3.0.1
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
wsproto==1.2.0
